Move mappings to separate file
